refactor(screenshots): extract seed layer from spec into playwright/seed/
Move all user provisioning, Talk room/message creation, calendar event
seeding, and note-to-self / reminder seeding out of the conversations
spec beforeAll and into a dedicated playwright/seed/ module.

global-setup.ts now calls seed() (all API seeding) then launches a
browser to capture storageState in playwright/.auth/state.json so each
test starts pre-authenticated without per-test login calls. Removes the
authCookies / beforeEach pattern from the spec.

AI-Assisted-By: Claude Sonnet 4.6 <noreply@anthropic.com>
