+++ playwright/global-setup.ts
@@ -2,7 +2,12 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { configureNextcloud, runOcc, startNextcloud, waitOnNextcloud } from '@nextcloud/e2e-test-server/docker'
+import { login } from '@nextcloud/e2e-test-server/playwright'
+import { User } from '@nextcloud/e2e-test-server'
+import { chromium } from '@playwright/test'
+import * as path from 'path'
 import { SCREENSHOT_PORT } from '../playwright.config'
+import { seed, seedNoteToSelf } from './seed'
 
 const SCREENSHOT_APPS = [
 	'activity',
@@ -16,6 +21,8 @@ const SCREENSHOT_APPS = [
 	'tasks',
 	'viewer',
 ]
+
+const AUTH_FILE = path.join(__dirname, '.auth', 'state.json')
 
 export default async function globalSetup() {
 	await startNextcloud('stable33', false, { exposePort: SCREENSHOT_PORT })
@@ -34,4 +41,25 @@ export default async function globalSetup() {
 	// Talk hides the "Message expiration" setting in conversation settings unless
 	// background jobs are in cron mode. Set it so the feature appears in the UI.
 	await runOcc(['config:app:set', 'core', 'backgroundjobs_mode', '--value', 'cron'])
+
+	// Seed all users and Talk data via API.
+	// seedTalk() returns the "Event planning" group token needed for post-browser seeding.
+	const eventToken = await seed()
+
+	// Launch a browser to initialise Talk (note-to-self requires a prior browser
+	// visit to /apps/spreed), seed post-browser data, then capture storageState so
+	// every test starts pre-authenticated without calling login() in beforeEach.
+	const browser = await chromium.launch()
+	const context = await browser.newContext({
+		baseURL: `http://localhost:${SCREENSHOT_PORT}`,
+		userAgent: 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/142.0.0.0 Safari/537.36',
+	})
+	const page = await context.newPage()
+	const christine = new User('christine', 'christine')
+	await login(page.request, christine)
+	await page.goto('/apps/spreed')
+	await page.locator('[aria-label="Conversation list"]').waitFor({ state: 'visible', timeout: 20000 }).catch(() => {})
+	await seedNoteToSelf(eventToken)
+	await context.storageState({ path: AUTH_FILE })
+	await browser.close()
 }
